refactor: isolate onboarding credential helpers
